requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
celery>=5.3.0
redis>=4.5.0
appium-python-client>=3.0.0
selenium>=4.0.0